Files edited to enable compilation
Most of the changes are if-else to switch conversions. Notable issues
with getopt (linker error if I use only the cosmopolitan version, so
some mix-n-match is happening) and xmalloc/xrealloc (make provides its
own version of these things).

A lot of tests pass. Failures usually involve -j and crash with fcntl
(returns EINVAL when I try to run multiple jobs).